+++ tests/test_mixin.py
@@ -114,21 +114,21 @@ class TestMeta(BaseTestCase):
         self.assertContains(response, 'itemscope itemtype="http://schema.org/Article"')
         self.assertContains(response, 'article:published_time"')
         self.assertContains(response, '<meta name="twitter:image:src" content="http://example.com/path/to/image">')
-        self.assertContains(response, '<link rel="author" href="https://plus.google.com/%s"/>' % meta.gplus_author)
-        self.assertContains(response, '<meta itemprop="description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta name="twitter:description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta property="og:description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta name="description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta name="keywords" content="%s">' % ", ".join(self.post.meta_keywords.split(",")))
+        self.assertContains(response, '<link rel="author" href="https://plus.google.com/{0}"/>'.format(meta.gplus_author))
+        self.assertContains(response, '<meta itemprop="description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta name="twitter:description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta property="og:description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta name="description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta name="keywords" content="{0}">'.format(', '.join(self.post.meta_keywords.split(","))))
 
     def test_templatetag_no_og(self):
         from meta import settings
         settings.USE_OG_PROPERTIES = False
         response = self.client.get('/title/')
         self.assertFalse(response.rendered_content.find('og:description') > -1)
-        self.assertContains(response, '<meta itemprop="description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta name="twitter:description" content="%s">' % self.post.meta_description)
-        self.assertContains(response, '<meta name="keywords" content="%s">' % ", ".join(self.post.meta_keywords.split(",")))
+        self.assertContains(response, '<meta itemprop="description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta name="twitter:description" content="{0}">'.format(self.post.meta_description))
+        self.assertContains(response, '<meta name="keywords" content="{0}">'.format(', '.join(self.post.meta_keywords.split(","))))
         settings.USE_OG_PROPERTIES = True
 
     def test_generic_prop_basically_works(self):